fix/rename-areas - rename temp areas
